[ add ] loading state & tsx => TypeScript
